Updated CMakeLists.txt for bounce_ball.cpp
If we reran CMake while we were figuring out the bounce_ball.cpp as a
group, it wouldn't have done anything.

Fixed it so that CMake would've worked if I chose to use it instead of
compiling by command line. Be sure to update the CMakeLists.txt when you
add new source files, and keep executables out of the src folder.