Adding a basic Movement Test in SDL
Added a WASD-based program that responds to user key presses by moving a very basic player sprite. This was done mainly to learn how to use the SDL2 library. Please note: the png files added are extremely basic. Feel free to modify them, make them more aesthetically pleasing, or scrap them altogether. This test was committed primarily to test the movement capability of objects in SDL. Finally, a Makefile and a symbolic link were added to the src directory for convenience.