SDL Test - Decomposing background image
Began to decompose the test background image into components, namely the small blue tent and the building, which were added as separate .png files. In addition, comments and pseudocode was written in the .cpp file to consider how we would generalize the code to work for any background, in any setting whatsoever. Please look over these comments, and we can share ideas and feedback soon.