A Revised General Engine, that can Run (though imperfectly)
Revised the general_engine program, which now runs without error (again, remember to press x to exit the window when running the executable). Also added a font generating test, which provides a functionality that could be useful in generating text on screen at a later time. Work still needs to be done in generalizing the background generation and proper determination of where the player is allowed to move. I will continue working through these problems (more details provided in comments).