Adding Linked List Implementation
Added a more general linked list implementation for the master_entity_list. This uses dynamic memory allocation for structures and thus will allow for an indefinite number of objects existing and being renderred on the screen at a time. Also added a makefile for those interested. Next steps are to update the map_matrix (2d array) more accurately using data from each entity. I also need to decompose the background PNG a little further in order to break down the well and perhaps even the lake in the background. Another update should be coming in the next week or two.