Updates to the main renderring engine
Adding an alternate (more likely to be used) background image. Also created a new directory just to develop the engine itself. More specifically, the new .cpp file engine_bst_woods_enemies.cpp is the newest version, and it (along with the new engine_bst.cpp) processes, in a rudimentary way, restrictions on where the player can move. It does this by reading data from a file about each entity's base coordinates and then updating the map matrix with these restrictions. Note: these newest versions of the engine (engine_bst.cpp and engine_bst_woods_enemies.cpp) compile properly when the stack protector is *disabled*. This means the flag -fno-stack-protector is used to compile them. If the stack protector IS used, then there may be a SigAbrt error. I have not found the cause of the error yet, but I will continue looking. Please let me know if you find it, or if we should simply use -fno-stack-protector at least in the short term.