This project is alive. (Added animations in linked_list implementations.)
Revised and updated the engine_linkedList.cpp implementation. It is now capable of handling some basic animations. (If you execute recent copies of the linkedList implementations, you'll notice that the bats flicker in the created window, and the player can still move as intended.) In the more recent copy, engine_ll_spaces.cpp, the code was made more readable, and some documentation was added. The linked-list implementations make the rendering order most well-defined and alterable. (Barring objections, this implementation will be used/maintained in the near future.)