Health bar loading and some new artwork
The engine_heap.cpp has been revised to be able to print, accurately, the player'd HP in the general case. (I recommend you have a look at the implementation, as there may be a more efficient way to deal with this.) However, the solution is general and should therefore work regardless of the value of the player HP. A ratio of 0.6 (6 out of 10) is used as an example; feel free to adjust this. Also see the new Actor class that has been added in engine_heap.hpp. The good news is, the project is far from dead.