More general background loading.
Refactored the code, in order to allow the engine to process a change in the background (e.g. when the character moves to a different screen). Also added more clear documentation. Will likely proceed to work on inventory strucutre next. What is the status of this project? It's alive and well.