Background change when walking to the left; updated README; etc
Added very basic inventory implementation (using an array - inventory limited by number of items). Press e to open the inventory. Items are added intermittently to the inventory as the game progresses over time (this is deliberate; it acts as a test on the inventory system). Also added capability of updating the background when walking off screen to the left. Updated the README. Next steps are to implement other systems (e.g. weight-based inventory, time-step system in the main loop, etc.).