Added more accurate time printing etc
Un-commented a line about clearing the renderer. In addition, the system can now print the number of microseconds elapsed per iteration of the game loop. This could be useful once we start getting moving actors (e.g. enemies) on the screen, as we'll need to reconcile the physics-update rate with the rendering framerate of the game. We should discuss this more.